requirements.txt
# MQTT client
paho-mqtt

# Camera library for Raspberry Pi
picamera2

# System metrics
psutil>=5.9.0

# OBC state machine
transitions

# Configuration
python-dotenv
pyyaml

# Sensor I2C libraries
smbus2
RPi.GPIO
lgpio

# Remote telemetry API
requests
